requirements.txt
mitmproxy
fastapi
uvicorn
pydantic
websocket-client
aiofiles
